chore(#3119): Improve pipeline editor component structure
