+++ ui/src/app/editor/components/pipeline-assembly/pipeline-assembly-options/pipeline-assembly-options-pipeline-cache/pipeline-assembly-options-pipeline-cache.component.html
@@ -16,21 +16,31 @@
   ~
   -->
 
-<div class="outer-assembly-preview" fxFlex="100">
-    <div class="outerAssembly" fxFlex="100">
-        <sp-pipeline-assembly-drawing-area
-            *ngIf="rawPipelineModel"
-            fxFlex="100"
-            style="position: relative"
-            [jsplumbBridge]="jsPlumbBridge"
-            [metricsInfo]="metricsInfo"
-            [rawPipelineModel]="rawPipelineModel"
-            [pipelineCanvasMetadata]="pipelineCanvasMetadata"
-            [pipelineCanvasMetadataAvailable]="
-                pipelineCanvasMetadata !== undefined
-            "
-            [readonly]="true"
-        ></sp-pipeline-assembly-drawing-area>
+<div
+    fxLayout="column"
+    fxLayoutAlign="start center"
+    class="pipeline-cache-block"
+>
+    <div
+        fxFlex="100"
+        fxLayoutAlign="start center"
+        fxLayout="row"
+        *ngIf="pipelineCached || pipelineCacheRunning"
+    >
+        <div fxLayout="row" fxLayoutAlign="start center">
+            <mat-spinner
+                [mode]="'indeterminate'"
+                color="accent"
+                [diameter]="15"
+                *ngIf="pipelineCacheRunning"
+            ></mat-spinner>
+            <span
+                >&nbsp;{{
+                    pipelineCacheRunning
+                        ? '&nbsp;Saving pipeline modifications'
+                        : 'All pipeline modifications saved.'
+                }}</span
+            >
+        </div>
     </div>
-    <ng-content></ng-content>
 </div>
