+++ ui/src/app/editor/components/pipeline-assembly/pipeline-assembly-options/pipeline-assembly-options-pipeline-cache/pipeline-assembly-options-pipeline-cache.component.scss
@@ -1,4 +1,4 @@
-/*
+/*!
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -15,3 +15,12 @@
  * limitations under the License.
  *
  */
+
+.pipeline-cache-progress {
+    display: inline-block;
+}
+
+.pipeline-cache-block {
+    height: 100%;
+    margin-left: 15px;
+}
