+++ ui/src/app/editor/components/pipeline-assembly/pipeline-assembly-options/pipeline-assembly-options.component.scss
@@ -1,4 +1,4 @@
-/*
+/*!
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -16,29 +16,10 @@
  *
  */
 
-@import '../../../../scss/_variables.scss';
-
-.mat-spinner-color::ng-deep svg circle {
-    stroke: var(--color-accent) !important;
-}
-
-.pipeline-assembly-options {
-    padding-left: 5px;
-    border: 1px solid var(--color-bg-3);
-}
-
-.outerAssembly {
-    left: 0;
-    top: 0;
-    display: flex;
-    flex: 1 1 0;
-    overflow: hidden;
-}
-
-.assembly-border {
-    border: 1px solid var(--color-bg-3);
-}
-
-.color-warn {
-    color: var(--color-warn);
+.assembly-options-divider {
+    width: 1px;
+    height: 70%;
+    margin-left: 10px;
+    margin-right: 10px;
+    background: var(--color-bg-3);
 }
