+++ ui/src/app/editor/components/pipeline-assembly/pipeline-assembly.component.ts
@@ -16,96 +16,58 @@
  *
  */
 
-import {
-    AfterViewInit,
-    Component,
-    ElementRef,
-    EventEmitter,
-    Input,
-    NgZone,
-    OnDestroy,
-    OnInit,
-    Output,
-    ViewChild,
-} from '@angular/core';
+import { AfterViewInit, Component, Input, ViewChild } from '@angular/core';
 import { JsplumbBridge } from '../../services/jsplumb-bridge.service';
 import { PipelinePositioningService } from '../../services/pipeline-positioning.service';
 import { PipelineValidationService } from '../../services/pipeline-validation.service';
-import { JsplumbService } from '../../services/jsplumb.service';
 import {
     PipelineElementConfig,
     PipelineElementUnion,
 } from '../../model/editor.model';
 import { ObjectProvider } from '../../services/object-provider.service';
-import {
-    ConfirmDialogComponent,
-    DialogService,
-    PanelType,
-    SpBreadcrumbService,
-} from '@streampipes/shared-ui';
+import { DialogService, PanelType } from '@streampipes/shared-ui';
 import { SavePipelineComponent } from '../../dialog/save-pipeline/save-pipeline.component';
-import { MatDialog } from '@angular/material/dialog';
 import { EditorService } from '../../services/editor.service';
 import {
     Pipeline,
     PipelineCanvasMetadata,
-    PipelineCanvasMetadataService,
-    PipelineService,
 } from '@streampipes/platform-services';
 import { JsplumbFactoryService } from '../../services/jsplumb-factory.service';
-import Panzoom, { PanzoomObject } from '@panzoom/panzoom';
-import { PipelineElementDraggedService } from '../../services/pipeline-element-dragged.service';
-import { PipelineComponent } from '../pipeline/pipeline.component';
-import { forkJoin, of, Subscription } from 'rxjs';
-import { PipelineElementDiscoveryComponent } from '../../dialog/pipeline-element-discovery/pipeline-element-discovery.component';
-import { SpPipelineRoutes } from '../../../pipelines/pipelines.routes';
+import { forkJoin } from 'rxjs';
 import { Router } from '@angular/router';
-import { catchError } from 'rxjs/operators';
+import { PipelineAssemblyDrawingAreaComponent } from './pipeline-assembly-drawing-area/pipeline-assembly-drawing-area.component';
+import { PipelineAssemblyOptionsComponent } from './pipeline-assembly-options/pipeline-assembly-options.component';
 
 @Component({
     selector: 'sp-pipeline-assembly',
     templateUrl: './pipeline-assembly.component.html',
     styleUrls: ['./pipeline-assembly.component.scss'],
 })
-export class PipelineAssemblyComponent
-    implements OnInit, AfterViewInit, OnDestroy
-{
+export class PipelineAssemblyComponent implements AfterViewInit {
     @Input()
     rawPipelineModel: PipelineElementConfig[];
 
     @Input()
-    currentModifiedPipelineId: string;
+    originalPipeline: Pipeline;
+
+    @Input()
+    pipelineCanvasMetadata: PipelineCanvasMetadata;
+
+    @Input()
+    pipelineCanvasMetadataAvailable = false;
 
     @Input()
     allElements: PipelineElementUnion[];
 
-    @Output()
-    pipelineCanvasMaximizedEmitter: EventEmitter<boolean> =
-        new EventEmitter<boolean>();
+    previewModeActive = false;
+    readonly: boolean;
 
     JsplumbBridge: JsplumbBridge;
-    currentZoomLevel: any;
-    preview: any;
 
-    selectMode: any;
-    originalPipeline: Pipeline;
-    pipelineValid = false;
-
-    pipelineCacheRunning = false;
-    pipelineCached = false;
-
-    pipelineCanvasMetadata: PipelineCanvasMetadata =
-        new PipelineCanvasMetadata();
-    pipelineCanvasMetadataAvailable = false;
-
-    config: any = {};
-    @ViewChild('outerCanvas') pipelineCanvas: ElementRef;
-
-    @ViewChild('pipelineComponent')
-    pipelineComponent: PipelineComponent;
-
-    panzoom: PanzoomObject;
-    moveSub: Subscription;
+    @ViewChild('assemblyOptionsComponent')
+    assemblyOptionsComponent: PipelineAssemblyOptionsComponent;
+    @ViewChild('drawingAreaComponent')
+    drawingAreaComponent: PipelineAssemblyDrawingAreaComponent;
 
     constructor(
         private jsPlumbFactoryService: JsplumbFactoryService,
@@ -113,125 +75,34 @@ export class PipelineAssemblyComponent
         private objectProvider: ObjectProvider,
         public editorService: EditorService,
         public pipelineValidationService: PipelineValidationService,
-        private pipelineService: PipelineService,
-        private jsplumbService: JsplumbService,
         private dialogService: DialogService,
-        private dialog: MatDialog,
-        private ngZone: NgZone,
         private router: Router,
-        private pipelineElementDraggedService: PipelineElementDraggedService,
-        private pipelineCanvasMetadataService: PipelineCanvasMetadataService,
-        private breadcrumbService: SpBreadcrumbService,
-    ) {
-        this.selectMode = true;
-        this.currentZoomLevel = 1;
-    }
-
-    ngOnInit(): void {
-        if (this.currentModifiedPipelineId) {
-            this.displayPipelineById(this.currentModifiedPipelineId);
-        } else {
-            this.checkAndDisplayCachedPipeline();
-        }
-        this.moveSub =
-            this.pipelineElementDraggedService.pipelineElementMovedSubject.subscribe(
-                position => {
-                    const offsetHeight =
-                        this.pipelineCanvas.nativeElement.offsetHeight;
-                    const offsetWidth =
-                        this.pipelineCanvas.nativeElement.offsetWidth;
-                    const currentPan = this.panzoom.getPan();
-                    let xOffset = 0;
-                    let yOffset = 0;
-                    if (position.y + currentPan.y > offsetHeight - 100) {
-                        yOffset = -10;
-                    }
-                    if (position.x + currentPan.x > offsetWidth - 100) {
-                        xOffset = -10;
-                    }
-                    if (xOffset < 0 || yOffset < 0) {
-                        this.pan(xOffset, yOffset);
-                    }
-                },
-            );
-    }
+    ) {}
 
     ngAfterViewInit() {
         this.JsplumbBridge = this.jsPlumbFactoryService.getJsplumbBridge(
-            this.preview,
+            this.readonly,
         );
-        const elem = document.getElementById('assembly');
-        this.panzoom = Panzoom(elem, {
-            maxScale: 5,
-            excludeClass: 'sp-no-pan',
-            canvas: true,
-            contain: 'outside',
-        });
-    }
-
-    autoLayout() {
-        this.pipelinePositioningService.layoutGraph(
-            '#assembly',
-            "div[id^='jsplumb']",
-            110,
-            false,
-        );
-        this.JsplumbBridge.repaintEverything();
-    }
-
-    zoomOut() {
-        this.doZoom(true);
-    }
-
-    zoomIn() {
-        this.doZoom(false);
-    }
-
-    doZoom(zoomOut) {
-        zoomOut ? this.panzoom.zoomOut() : this.panzoom.zoomIn();
-        this.currentZoomLevel = this.panzoom.getScale();
-        this.JsplumbBridge.setZoom(this.currentZoomLevel);
-        this.JsplumbBridge.repaintEverything();
-    }
-
-    showClearAssemblyConfirmDialog(event: any) {
-        const dialogRef = this.dialog.open(ConfirmDialogComponent, {
-            width: '500px',
-            data: {
-                title: 'Do you really want to delete the current pipeline?',
-                subtitle: 'This cannot be undone.',
-                cancelTitle: 'No',
-                okTitle: 'Yes',
-                confirmAndCancel: true,
-            },
-        });
-        dialogRef.afterClosed().subscribe(ev => {
-            if (ev) {
-                if (this.currentModifiedPipelineId) {
-                    this.currentModifiedPipelineId = undefined;
-                }
-                this.clearAssembly();
-                this.editorService.makePipelineAssemblyEmpty(true);
-            }
-        });
     }
 
     /**
      * clears the Assembly of all elements
      */
     clearAssembly() {
+        this.editorService.makePipelineAssemblyEmpty(true);
         this.JsplumbBridge.deleteEveryEndpoint();
         this.rawPipelineModel = [];
-        this.currentZoomLevel = 1;
-        this.JsplumbBridge.setZoom(this.currentZoomLevel);
+        this.drawingAreaComponent.resetZoom();
         this.JsplumbBridge.repaintEverything();
 
         forkJoin([
             this.editorService.removePipelineFromCache(),
             this.editorService.removeCanvasMetadataFromCache(),
-        ]).subscribe(msg => {
-            this.pipelineCached = false;
-            this.pipelineCacheRunning = false;
+        ]).subscribe(() => {
+            this.pipelineCanvasMetadata = new PipelineCanvasMetadata();
+            if (this.originalPipeline) {
+                this.router.navigate(['pipelines', 'create']);
+            }
         });
     }
 
@@ -239,7 +110,8 @@ export class PipelineAssemblyComponent
      * Sends the pipeline to the server
      */
     submit() {
-        const pipelineModel = this.pipelineComponent.rawPipelineModel;
+        //const pipelineModel = this.pipelineComponent.rawPipelineModel;
+        const pipelineModel = this.rawPipelineModel;
         const pipeline = this.objectProvider.makePipeline(pipelineModel);
         this.pipelinePositioningService.collectPipelineElementPositions(
             this.pipelineCanvasMetadata,
@@ -247,7 +119,7 @@ export class PipelineAssemblyComponent
         );
         pipeline.valid = this.pipelineValidationService.isValidPipeline(
             pipelineModel,
-            this.preview,
+            this.readonly,
         );
         const dialogRef = this.dialogService.open(SavePipelineComponent, {
             panelType: PanelType.SLIDE_IN_PANEL,
@@ -265,9 +137,7 @@ export class PipelineAssemblyComponent
             .subscribe((config: { reload: boolean; pipelineId: string }) => {
                 if (config?.reload) {
                     this.clearAssembly();
-                    this.editorService.makePipelineAssemblyEmpty(true);
                     this.rawPipelineModel = [];
-                    this.currentModifiedPipelineId = undefined;
                     setTimeout(() => {
                         this.router.navigate(['pipelines', 'create']);
                     });
@@ -275,149 +145,12 @@ export class PipelineAssemblyComponent
             });
     }
 
-    checkAndDisplayCachedPipeline() {
-        const cachedPipeline = this.editorService.getCachedPipeline();
-        const cachedCanvasMetadata =
-            this.editorService.getCachedPipelineCanvasMetadata();
-        forkJoin([cachedPipeline, cachedCanvasMetadata]).subscribe(results => {
-            if (results[0] && results[0].length > 0) {
-                this.rawPipelineModel = results[0] as PipelineElementConfig[];
-                this.handleCanvasMetadataResponse(results[1]);
-                this.displayPipelineInEditor(
-                    !this.pipelineCanvasMetadataAvailable,
-                    this.pipelineCanvasMetadata,
-                );
-            }
-        });
+    togglePreview(): void {
+        this.previewModeActive = !this.previewModeActive;
+        this.drawingAreaComponent.initiatePipelineElementPreview();
     }
 
-    displayPipelineById(pipelineId: string) {
-        const pipelineReq = this.pipelineService.getPipelineById(pipelineId);
-        const canvasMetadataReq = this.pipelineCanvasMetadataService
-            .getPipelineCanvasMetadata(pipelineId)
-            .pipe(
-                catchError(error => {
-                    this.handleCanvasMetadataResponse(undefined);
-                    return of(undefined);
-                }),
-            );
-
-        forkJoin([pipelineReq, canvasMetadataReq]).subscribe(
-            ([pipelineResp, canvasResp]) => {
-                if (pipelineResp) {
-                    this.originalPipeline = pipelineResp;
-                    this.breadcrumbService.updateBreadcrumb([
-                        SpPipelineRoutes.BASE,
-                        { label: this.originalPipeline.name },
-                        { label: 'Modify' },
-                    ]);
-                    this.rawPipelineModel = this.jsplumbService.makeRawPipeline(
-                        this.originalPipeline,
-                        false,
-                    );
-                }
-                if (canvasResp !== undefined) {
-                    this.handleCanvasMetadataResponse(canvasResp);
-                }
-                this.displayPipelineInEditor(
-                    !this.pipelineCanvasMetadataAvailable,
-                    this.pipelineCanvasMetadata,
-                );
-            },
-        );
-    }
-
-    handleCanvasMetadataResponse(canvasMetadata: PipelineCanvasMetadata) {
-        if (canvasMetadata) {
-            this.pipelineCanvasMetadata = canvasMetadata;
-            this.pipelineCanvasMetadataAvailable = true;
-        } else {
-            this.pipelineCanvasMetadataAvailable = false;
-            this.pipelineCanvasMetadata = new PipelineCanvasMetadata();
-        }
-    }
-
-    displayPipelineInEditor(
-        autoLayout: boolean,
-        pipelineCanvasMetadata?: PipelineCanvasMetadata,
-    ): void {
-        setTimeout(() => {
-            this.pipelinePositioningService.displayPipeline(
-                this.rawPipelineModel,
-                '#assembly',
-                false,
-                autoLayout,
-                pipelineCanvasMetadata,
-            );
-            this.editorService.makePipelineAssemblyEmpty(false);
-            this.ngZone.run(() => {
-                this.pipelineValid =
-                    this.pipelineValidationService.isValidPipeline(
-                        this.rawPipelineModel.filter(
-                            pe => !pe.settings.disabled,
-                        ),
-                        false,
-                    );
-            });
-            this.pipelineComponent.triggerPipelineModification();
-        });
-    }
-
-    isPipelineAssemblyEmpty() {
-        return (
-            this.rawPipelineModel.length === 0 ||
-            this.rawPipelineModel.every(pe => pe.settings.disabled)
-        );
-    }
-
-    panLeft() {
-        this.pan(100, 0);
-    }
-
-    panRight() {
-        this.pan(-100, 0);
-    }
-
-    panUp() {
-        this.pan(0, 100);
-    }
-
-    panDown() {
-        this.pan(0, -100);
-    }
-
-    panHome() {
-        this.panAbsolute(0, 0);
-    }
-
-    pan(xOffset: number, yOffset: number) {
-        const currentPan = this.panzoom.getPan();
-        const panX = Math.min(0, currentPan.x + xOffset);
-        const panY = Math.min(0, currentPan.y + yOffset);
-        this.panzoom.pan(panX, panY);
-    }
-
-    panAbsolute(x: number, y: number) {
-        this.panzoom.pan(x, y);
-    }
-
-    triggerPipelinePreview() {
-        this.pipelineComponent.initiatePipelineElementPreview();
-    }
-
-    openDiscoverDialog() {
-        this.dialogService.open(PipelineElementDiscoveryComponent, {
-            panelType: PanelType.SLIDE_IN_PANEL,
-            title: 'Discover pipeline elements',
-            width: '50vw',
-            data: {
-                currentElements: this.allElements,
-                rawPipelineModel: this.rawPipelineModel,
-            },
-        });
-    }
-
-    ngOnDestroy() {
-        this.moveSub?.unsubscribe();
+    triggerCacheUpdate(): void {
+        this.assemblyOptionsComponent.triggerCacheUpdate();
     }
 }
