+++ ui/src/app/editor/services/jsplumb-factory.service.ts
@@ -30,10 +30,7 @@ import { JsplumbConfigService } from './jsplumb-config.service';
 @Injectable({ providedIn: 'root' })
 export class JsplumbFactoryService {
     pipelineEditorInstance: BrowserJsPlumbInstance;
-    pipelinePreviewInstance: BrowserJsPlumbInstance;
-
     pipelineEditorBridge: JsplumbBridge;
-    pipelinePreviewBridge: JsplumbBridge;
 
     constructor(
         private pipelineElementDraggedService: PipelineElementDraggedService,
@@ -41,26 +38,16 @@ export class JsplumbFactoryService {
     ) {}
 
     getJsplumbBridge(previewConfig: boolean): JsplumbBridge {
-        if (!previewConfig) {
-            if (!this.pipelineEditorBridge) {
-                this.pipelineEditorInstance = this.makePipelineEditorInstance();
-                this.prepareJsplumb(this.pipelineEditorInstance);
-                this.pipelineEditorBridge = new JsplumbBridge(
-                    this.pipelineEditorInstance,
-                );
-            }
-            return this.pipelineEditorBridge;
-        } else {
-            if (!this.pipelinePreviewBridge) {
-                this.pipelinePreviewInstance =
-                    this.makePipelinePreviewInstance();
-                this.prepareJsplumb(this.pipelinePreviewInstance);
-                this.pipelinePreviewBridge = new JsplumbBridge(
-                    this.pipelinePreviewInstance,
-                );
-            }
-            return this.pipelinePreviewBridge;
+        if (!this.pipelineEditorBridge) {
+            this.pipelineEditorInstance = previewConfig
+                ? this.makePipelinePreviewInstance()
+                : this.makePipelineEditorInstance();
+            this.prepareJsplumb(this.pipelineEditorInstance);
+            this.pipelineEditorBridge = new JsplumbBridge(
+                this.pipelineEditorInstance,
+            );
         }
+        return this.pipelineEditorBridge;
     }
 
     makePipelineEditorInstance(): BrowserJsPlumbInstance {
@@ -82,7 +69,7 @@ export class JsplumbFactoryService {
 
     makePipelinePreviewInstance(): BrowserJsPlumbInstance {
         return newInstance({
-            container: document.getElementById('assembly-preview'),
+            container: document.getElementById('assembly'),
             elementsDraggable: false,
         });
     }
@@ -93,13 +80,8 @@ export class JsplumbFactoryService {
         );
     }
 
-    destroy(preview: boolean) {
-        if (preview) {
-            this.pipelinePreviewInstance.destroy();
-            this.pipelinePreviewBridge = undefined;
-        } else {
-            this.pipelineEditorInstance.destroy();
-            this.pipelineEditorBridge = undefined;
-        }
+    destroy() {
+        this.pipelineEditorInstance?.destroy();
+        this.pipelineEditorBridge = undefined;
     }
 }
