+++ ui/src/app/pipeline-details/components/pipeline-details-expansion-panel/pipeline-element-details-row/elements/pipeline-elements-row.component.ts
@@ -24,7 +24,7 @@ import { PipelineElementTypeUtils } from '../../../../../editor/utils/editor.uti
     selector: 'sp-pipeline-elements-row',
     templateUrl: './pipeline-elements-row.component.html',
     styleUrls: [
-        '../../../../../editor/components/pipeline-element-icon-stand-row/pipeline-element-icon-stand-row.component.scss',
+        '../../../../../editor/components/pipeline-element-icon-stand/pipeline-element-icon-stand-row/pipeline-element-icon-stand-row.component.scss',
     ],
 })
 export class PipelineElementsRowComponent implements OnInit {
