+++ ui/src/app/pipeline-details/pipeline-details.component.html
@@ -32,7 +32,6 @@
     </div>
     <div fxFlex="100" fxLayout="column">
         <sp-pipeline-preview
-            [jspcanvas]="'assembly-preview'"
             [metricsInfo]="metricsInfo"
             [pipeline]="pipeline"
             [pipelineCanvasMetadata]="pipelineCanvasMetadata"
